added authentication and locked up my utations and queries

+++ src/resolvers/Mutation.js
@@ -1,29 +1,72 @@
+import bcrypt from "bcryptjs";
+import jwt from "jsonwebtoken";
+import getUserId from "../utils/getUserId";
+
 const Mutation = {
-  async createUser(parent, args, { prisma }, info) {
-    return prisma.mutation.createUser({ data: args.data }, info);
+  async login(parent, args, { prisma }, info) {
+    const user = await prisma.query.user({ where: { email: args.data.email } });
+
+    if (!user) {
+      throw new Error("Unable to log in");
+    }
+
+    const isMatch = bcrypt.compare(args.data.password, user.password);
+
+    if (!isMatch) {
+      throw new Error("Incorrect password");
+    }
+
+    const token = jwt.sign({ userId: user.id }, "thisisasecret");
+
+    return { user, token };
   },
-  async deleteUser(parent, args, { prisma }, info) {
+  async createUser(parent, args, { prisma }, info) {
+    if (args.data.password.length < 8) {
+      throw new Error("Password must be 8 characters or longer");
+    }
+
+    const password = await bcrypt.hash(args.data.password, 10);
+
+    const user = await prisma.mutation.createUser({
+      data: {
+        ...args.data,
+        password
+      }
+    });
+
+    return {
+      user,
+      token: jwt.sign({ userId: user.id }, "thisisasecret")
+    };
+  },
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.deleteUser(
       {
         where: {
-          id: args.id
+          id: userId
         }
       },
       info
     );
   },
-  async updateUser(parent, args, { prisma }, info) {
+  async updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.updateUser(
       {
         where: {
-          id: args.id
+          id: userId
         },
         data: args.data
       },
       info
     );
   },
-  createPost(parent, args, { prisma }, info) {
+  createPost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.createPost(
       {
         data: {
@@ -32,7 +75,7 @@ const Mutation = {
           published: args.data.published,
           author: {
             connect: {
-              id: args.data.author
+              id: userId
             }
           }
         }
@@ -40,7 +83,20 @@ const Mutation = {
       info
     );
   },
-  deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!postExists) {
+      throw new Error("Unable to delete post");
+    }
+
     return prisma.mutation.deletePost(
       {
         where: {
@@ -50,7 +106,20 @@ const Mutation = {
       info
     );
   },
-  updatePost(parent, args, { prisma }, info) {
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!postExists) {
+      throw new Error("Unable to delete post");
+    }
+
     return prisma.mutation.updatePost(
       {
         where: {
@@ -62,13 +131,15 @@ const Mutation = {
     );
   },
   createComment(parent, args, { prisma }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.createComment(
       {
         data: {
           text: args.data.text,
           author: {
             connect: {
-              id: args.data.author
+              id: userId
             }
           },
           post: {
@@ -81,7 +152,20 @@ const Mutation = {
       info
     );
   },
-  deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma }, info) {
+    const userId = getUserId(request);
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!commentExists) {
+      throw new Error("Unable to delete comment");
+    }
+
     return prisma.mutation.deleteComment(
       {
         where: {
@@ -91,7 +175,20 @@ const Mutation = {
       info
     );
   },
-  updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma }, info) {
+    const userId = getUserId(request);
+
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!commentExists) {
+      throw new Error("Unable to delete comment");
+    }
+
     return prisma.mutation.updateComment(
       {
         where: {
